adding feed testhelper to separate feed tests from target types (currently only worksheet)
